R6.4: split vector/lower into program/signal/tables/check
Facade mod.rs re-exports everything; no public path changes. Program
artifacts (region bodies, VerifiedPureVectorProgram, lowering context,
error taxonomy) live in program.rs; the producer lowerer and program
entry points in signal.rs; shared table-name/table-signal helpers in
tables.rs; boundary and body verification in check.rs.

Plan §4.8 note: lower has no standalone checker entry point —
verify_plan_prepared_boundary and verify_pure_vector_bodies are called
only from the producer's terminal step (lower_vector_program_impl).
No admission guard changed paths in this split; the checks moved intact
into check.rs with the same single call site.

mutable_table_name is re-exported explicitly at
pub(in crate::signal_fir::vector) because vector/module imports it via
super::lower::.

Validation: transform 387 lib tests, transform clippy (+1.97.0) clean.
