+++ porting/journal/2026-07-20.md
@@ -88,3 +88,20 @@ modules on both sides of the plan artifact. Validation: transform 385,
   (= route) par le déplacement — élargi en
   `pub(in crate::signal_fir::vector)` car `lower` l'importe.
 - Gates : transform 387, clippy 0, fmt vert.
+
+## R6.4 — scission de `vector/lower` — R6 complet
+
+- `lower/mod.rs` (2 629 lignes) scindé en `program.rs` (259, artefacts +
+  erreurs), `signal.rs` (1 934, lowerer producteur + entrées de programme),
+  `tables.rs` (50, helpers de tables partagés), `check.rs` (383,
+  vérifications frontière/corps), façade `mod.rs` (29).
+- §4.8 (cas dégénéré) : `lower` n'a PAS d'entrée checker autonome —
+  `verify_plan_prepared_boundary` et `verify_pure_vector_bodies` ne sont
+  appelées que depuis l'étape terminale du producteur
+  (`lower_vector_program_impl`). Aucune garde n'a changé de chemin ; les
+  vérifications ont déménagé intactes vers `check.rs` avec le même site
+  d'appel unique. Rien à re-tester par une entrée checker qui n'existe pas.
+- `mutable_table_name` : re-export explicite
+  `pub(in crate::signal_fir::vector)` dans la façade (le module `module`
+  l'importe par `super::lower::`).
+- Gates : transform 387, clippy 0, fmt vert.
